==> notebook/4_regression-analysis.ipynb <==
%matplotlib inline

from IPython.display import display, HTML
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns
import sklearn.preprocessing
import statsmodels.api as sm
import statsmodels.formula.api as smf
from tqdm import tqdm_notebook as tqdm

from collections import Counter
import operator
import sys

def normalize(series):
    return sklearn.preprocessing.quantile_transform(
            series.values.reshape(-1, 1), copy=True)

authors_df = pd.read_csv("../data/processed/author-label-state.csv.bz2", index_col=0)

authors_df = authors_df.merge(pd.read_csv("../data/processed/states-extra.csv"), how='left')

authors_df.columns = [c.replace(' ', '') for c in authors_df.columns]
len(authors_df)

Counter(authors_df.label)

author2label = dict(authors_df[['author', 'label']].values)

graph_df = pd.read_csv(
    "../data/processed/parent_child_sentiment_edges_politics.csv.bz2", index_col=0)
graph_df.head()

graph_df['lparent'] = graph_df.parent.apply(author2label.get)
graph_df['lchild'] = graph_df.child.apply(author2label.get)

authors = set(graph_df.parent) | set(graph_df.child)
len(authors), len(authors) ** 2

edges = list(map(tuple, graph_df[['parent', 'child']].values))
len(edges)

Counter([author2label[c] for p, c in edges])

#edge-based sampling

np.random.seed(123)
heads, tails = np.array(edges).T
edges_set = set(edges)
non_edges = list()
num_non_edges = len(edges)
pbar = tqdm(total=num_non_edges)
while len(non_edges) < num_non_edges:
    sample_size = min(2 ** 17, num_non_edges - len(non_edges))
    head_samples = np.random.choice(heads, sample_size)
    tail_samples = np.random.choice(tails, sample_size)
    sample = map(tuple, np.array([head_samples, tail_samples]).T)
    new_non_edges = [e for e in sample if e not in edges_set]
    non_edges += new_non_edges
    pbar.update(len(new_non_edges))
pbar.close()
del edges_set
len(non_edges)


Counter([author2label[c] for p, c in non_edges])

matrix = np.array([
    (
        parent,
        child,
        int(author2label[child] == 'D'),
        int(author2label[parent] != author2label[child]),
        int(y)
    )
    for y, examples in [(True, edges), (False, non_edges)]
    for parent, child in tqdm(examples, desc=f"Evaluating {y} examples")
])

np.random.seed(123)
balanced_examples = pd.DataFrame(np.random.permutation(matrix), columns=[
    "parent", "child", "dem_author", "diff_label", "is_link"
])

balanced_examples['dem_author'] = balanced_examples.dem_author.astype(np.int64, copy=False)
balanced_examples['diff_label'] = balanced_examples.diff_label.astype(np.int64, copy=False)
balanced_examples['is_link'] = balanced_examples.is_link.astype(np.int64, copy=False)

examples = balanced_examples


examples[['dem_author', 'diff_label', 'is_link']].mean()

authors_df['clintonShare'] = authors_df.clintonVotes / authors_df.totalVotes
authors_df['trumpShare'] = authors_df.trumpVotes / authors_df.totalVotes
authors_df['nonvote'] = 1. - (authors_df.trumpVotes + authors_df.clintonVotes) / authors_df.totalVotes

authors_df['GiniCoefficient_q'] = normalize(authors_df['GiniCoefficient'])
authors_df['income_q'] = normalize(authors_df['income'])
authors_df['clintonShare_q'] = normalize(authors_df['clintonShare'])
authors_df['trumpShare_q'] = normalize(authors_df['trumpShare'])
authors_df['nonvote_q'] = normalize(authors_df['nonvote'])
authors_df['with_diploma_q'] = normalize(authors_df['with_diploma'])
authors_df['unemployment2016_q'] = normalize(authors_df['unemployment2016'])

scores = pd.read_csv(
        "../data/processed/scores-of-authors-in-politics-from-author-label-state.csv.bz2")

scores['frac_pos'] = scores.num_comm_pos / scores.num_comm

scores['num_comm_q'] = normalize(scores['num_comm'])
scores['avg_score_q'] = normalize(scores['avg_score'])
scores['frac_pos_q'] = normalize(scores['frac_pos'])

dataset = (examples
           .merge(authors_df, left_on='child', right_on='author', how='left')
           .drop(columns=['author'])
           .merge(authors_df[['author', 'state']].rename(columns={'state': 'state_parent'}),
                  left_on='parent', right_on='author', how='left')
           .drop(columns=['author'])
           .merge(scores[['author', 'num_comm_q', 'avg_score_q', 'frac_pos_q']],
               left_on='child', right_on='author', how='left')
           .merge(scores[['author', 'num_comm_q', 'avg_score_q', 'frac_pos_q']],
               left_on='parent', right_on='author', how='left', suffixes=['_c', '_p'])
           .drop(columns=['author_c', 'author_p', 'parent', 'child', 'label'])
)

dataset['same_state'] = (dataset.state == dataset.state_parent).astype(np.float64)
dataset = dataset.drop(columns=['state', 'state_parent'])

dataset['diff_avg_score_q'] = np.abs(dataset['avg_score_q_c'] - dataset['avg_score_q_p'])
dataset['diff_frac_pos_q'] = np.abs(dataset['frac_pos_q_c'] - dataset['frac_pos_q_p'])
dataset['diff_num_comm_q'] = np.abs(dataset['num_comm_q_c'] - dataset['num_comm_q_p'])


pvalue2asterisk = lambda x: '***' if x < 0.001 else '*' if x < 0.05 else ''
    
def add_results_to_table(res, df=None):
#     main_stat = pd.Series('{:.5f}'.format(res.prsquared), index=['Pseudo-$R^2$'])
    main_stat = pd.Series([res.aic], index=['AIC'])
    logits = np.exp(res.params)
#     logits = res.tvalues # Which actually is Z values
    r = pd.concat([res.pvalues, logits], axis=1)
    new_row = r.transpose().apply(lambda x: '{:.3f}'.format(x[1]) + pvalue2asterisk(x[0]))
    new_row = pd.concat([main_stat, new_row])
    if df is None:
        return pd.DataFrame([new_row])
    else:
        return df.append(new_row, ignore_index=True)

base_formula = "is_link ~ dem_author + diff_label + dem_author*diff_label"
res = smf.logit(formula=base_formula,
               data=dataset.astype(np.float64)).fit()

df = add_results_to_table(res)

coef_int, coef_dem, coef_diff, coef_dem_diff = res.params.values
matrix_index = pd.MultiIndex.from_tuples(
    [('C', 'C'), ('C', 'T'), ('T', 'C'), ('T', 'T')],
    names=['child', 'parent'])
base_matrix = np.exp(pd.DataFrame(index=matrix_index, columns=['Odds ratio'], data=[
    coef_int + coef_dem,
    coef_int + coef_dem + coef_diff + coef_dem_diff,
    coef_int + coef_diff,
    coef_int
]))
base_matrix

with open("../logit-matrix.tex", "w") as f:
    f.write(
r"""\bordermatrix{ & T & C \cr
      T & %.3f & %.3f \cr
      C & %.3f & %.3f  } 
""" % (base_matrix.loc[('T', 'T')], base_matrix.loc[('T', 'C')],
       base_matrix.loc[('C', 'T')], base_matrix.loc[('C', 'C')])    
)

other_reddit_variables = ["avg_score_q", "frac_pos_q"]
for variable in tqdm(other_reddit_variables):
    res = smf.logit(formula=base_formula + f" + {variable}_c + {variable}_p + diff_{variable}",
                   data=dataset.astype(np.float64)).fit()
#     display(res.summary())
    df = add_results_to_table(res, df)

full_reddit_variables = "".join([f" + {variable}_c + {variable}_p + diff_{variable}"
                           for variable in other_reddit_variables])

res = smf.logit(formula=base_formula + full_reddit_variables,
               data=dataset.astype(np.float64)).fit()
# display(res.summary())
df = add_results_to_table(res, df)

variable_rename = {
    'dem_author': 'Clinton sup.',
    'diff_label': 'Cross-group',
    'avg_score_q': 'Avg. score',
    'frac_pos_q': 'Frac. positive',
    'num_comm_q': 'No. comments',
    'same_state': 'Same state',
    'swing': 'Swing state',
    'clintonShare_q': 'Clinton share',
    'trumpShare_q': 'Trump share',
    'nonvote_q': 'Non-vote share',
    'unemployment2016_q': 'Unemployment',
    'GiniCoefficient_q': 'Gini Coefficient',
    'income_q': 'Median Income',
    'with_diploma_q': 'High school'
}

def rename(x):
    if ':' in x:
        return ', '.join(map(rename, x.split(':')))
    if x in variable_rename:
        return variable_rename[x]
    if x[:-2] in variable_rename:
        return variable_rename[x[:-2]] + (" (author)" if x[-2:] == '_c' else " (target)")
    if x[:5] == 'diff_':
        return "Diff. " + rename(x[5:]).lower()
    return x


base_AIC = df.AIC[0]

df['$\Delta$ AIC'] = base_AIC - df.AIC

table1 = df[df.columns[2:]].rename(columns={k: rename(k) for k in df.columns}).transpose().fillna('')
table1

with open("../logit-table-reddit.tex", "w") as f:
    latex = table1.to_latex(header=False, escape=False).replace(r"\toprule", r"""
\toprule
Variable name &  \multicolumn{%d}{c}{Odds ratios. * $p < 0.05$, *** $p < 0.001$} \\
\midrule
    """ % len(table1.columns)).replace("$\Delta$", r"\midrule $\Delta$")
    f.write(latex)

real_variables = [
     'same_state',
     'swing',
     'clintonShare_q',
     'trumpShare_q',
     'nonvote_q',
     'unemployment2016_q',
     'GiniCoefficient_q',
     'income_q',
     'with_diploma_q',
]

df = None
for variable in tqdm(real_variables):
    res = smf.logit(formula=(base_formula + full_reddit_variables +
                    f" + {variable} + {variable}*diff_label"),
               data=dataset.astype(np.float64)).fit()
    if variable == 'same_state':
        df = add_results_to_table(res, df)
    else:
        res = smf.logit(formula=(base_formula + full_reddit_variables +
                    f" + same_state + same_state*diff_label + {variable} + {variable}*diff_label"),
               data=dataset.astype(np.float64)).fit()
        df = add_results_to_table(res, df)

table2 = df[df.columns[2:]].rename(columns={k: rename(k) for k in df.columns}).transpose().fillna('')

with open("../logit-table-realworld.tex", "w") as f:
    latex = table2.to_latex(header=False, escape=False).replace(r"\toprule", r"""
\toprule
Variable name &  \multicolumn{%d}{c}{Odds ratios. * $p < 0.05$, *** $p < 0.001$} \\
\midrule
    """ % len(table2.columns)).replace("$\Delta$", r"\midrule $\Delta$")
    f.write(latex)

table2
